web-ahp-app/backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
numpy>=1.26.0,<3.0.0
scipy>=1.11.0,<2.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
